# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Bootcamp - OpenAPI 3.0
  description: Contrato de Banco Bootcamp
  version: 1.0.0
servers:
  - url: nttdata/
...
paths:
  /person:
    get:
      tags:
        - Person
      summary: Find All Customers Personal
      description: Find All Customers Personal
      operationId: findAllPerson
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponseDto'
    put:
      tags:
        - Person
      summary: Update an existing customer personal
      description: Update an existing customer personal by Id
      operationId: updatePerson
      requestBody:
        description: Update an existent customer in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponseDto'
    post:
      tags:
        - Person
      summary: Add a new customer personal to the bank
      description: Add a new customer personal to the bank
      operationId: createPerson
      requestBody:
        description: Create a new customer in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponseDto'
  /person/{personId}:
    get:
      tags:
        - Person
      summary: Find Customers Personal by id
      description: Find Customers Personal by id
      operationId: findPersonById
      parameters:
        - name: personId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponseDto'
...
  /transaction:
    get:
      tags:
        - Transaction
      summary: Find All Transaction
      description: Find All Transaction
      operationId: findAllTransaction
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseDto'
    put:
      tags:
        - Transaction
      summary: Update an existing transaction
      description: Update an existing transaction by Id
      operationId: updateTransaction
      requestBody:
        description: Update an existent transaction in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseDto'
    post:
      tags:
        - Transaction
      summary: Add a new transaction to the bank
      description: Add a new transaction to the bank
      operationId: createTransaction
      requestBody:
        description: Create a new customer company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseDto'
  /transaction/{transactionId}:
    get:
      tags:
        - Transaction
      summary: Find Transaction by id
      description: Find Transaction by id
      operationId: findTransactionById
      parameters:
        - name: transactionId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseDto'
  /transaction/customer/{customerId}:
    get:
      tags:
        - Transaction
      summary: Find Transaction by customerId
      description: Find Transaction by customerId
      operationId: findTransactionByCustomerId
      parameters:
        - name: customerId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseDto'
  /transaction/product/{productId}:
    get:
      tags:
        - Transaction
      summary: Find Transaction by productId
      description: Find Transaction by productId
      operationId: findTransactionByProductId
      parameters:
        - name: productId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseDto'
  /credit:
    get:
      tags:
        - Credit
      summary: Find All Credit
      description: Find All Credit
      operationId: findAllCredit
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
    put:
      tags:
        - Credit
      summary: Update an existing Credit
      description: Update an existing Credit by Id
      operationId: updateCredit
      requestBody:
        description: Update an existent Credit in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
  /credit/person:
    post:
      tags:
        - Credit
      summary: Add a new Credit person to the bank
      description: Add a new Credit person to the bank
      operationId: createCreditPerson
      requestBody:
        description: Create a new Credit person in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
  /credit/company:
    post:
      tags:
        - Credit
      summary: Add a new Credit company to the bank
      description: Add a new Credit company to the bank
      operationId: createCreditCompany
      requestBody:
        description: Create a new Credit company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
  /credit/{creditId}:
    get:
      tags:
        - Credit
      summary: Find Credit by id
      description: Find Credit by id
      operationId: findCreditById
      parameters:
        - name: creditId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
  /credit/pay:
    post:
      tags:
        - Credit
      summary: Pay a credit person or company to the bank
      description: Pay a credit person or company to the bank
      operationId: payCredit
      requestBody:
        description: Pay a Credit person or company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditPayRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditPayResponseDto'
...
components:
  schemas:
    Person:
      required:
        - dni
        - email
        - id
        - lastName
        - name
        - telephone
        - profile
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        name:
          type: string
          example: Antonio
        lastName:
          type: string
          example: Silva
        dni:
          type: string
          example: "76893214"
        email:
          type: string
          example: antsi@correo.com
        telephone:
          type: string
          example: "945856325"
        profile:
          type: string
          example: ESTANDAR
    PersonResponseDto:
      required:
        - id
        - name
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
        person:
          type: array
          items:
            $ref: '#/components/schemas/Person'
    PersonRequestDto:
      required:
        - dni
        - email
        - lastName
        - name
        - telephone
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        name:
          type: string
          example: Antonio
        lastName:
          type: string
          example: Silva
        dni:
          type: string
          example: "76893214"
        email:
          type: string
          example: antsi@correo.com
        telephone:
          type: string
          example: "945856325"
...
    MessageStatus:
      required:
        - message
        - statusCode
      type: object
      properties:
        statusCode:
          type: string
          example: "0"
        message:
          type: string
          example: Operacion exitosa.
...
    Transaction:
      required:
        - amount
        - customerId
        - id
        - productId
        - productType
        - transactionDate
        - transactionType
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        productType:
          type: string
          example: AHORRO
        productId:
          type: string
          example: "1"
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        transactionType:
          type: string
          example: RETIRO
        amount:
          type: number
          format: double
          example: 500
        transactionDate:
          type: string
          example: 12-03-2023
    TransactionResponseDto:
      required:
        - id
        - name
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
        transaction:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    TransactionRequestDto:
      required:
        - amount
        - customerId
        - productId
        - productType
        - transactionDate
        - transactionType
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        productType:
          type: string
          example: AHORRO
        productId:
          type: string
          example: "1"
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        transactionType:
          type: string
          example: RETIRO
        amount:
          type: number
          format: double
          example: 500
        transactionDate:
          type: string
          example: 12-03-2023
    Credit:
      required:
        - amountFee
        - createDate
        - customerId
        - customerType
        - id
        - payDay
        - quantityFee
        - settlementAmount
        - totalAmount
        - totalFee
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerType:
          type: string
          example: PERSONAL
        totalAmount:
          type: number
          format: double
          example: 500
        amountFee:
          type: number
          format: double
          example: 18
        quantityFee:
          type: integer
          example: 12
        totalFee:
          type: integer
          example: 12
        settlementAmount:
          type: number
          format: double
          example: 500
        payDay:
          type: string
          example: 12-03-2023
        createDate:
          type: string
          example: 12-03-2023
    CreditResponseDto:
      required:
        - credit
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
        credit:
          type: array
          items:
            $ref: '#/components/schemas/Credit'
    CreditRequestDto:
      required:
        - amountFee
        - createDate
        - customerId
        - customerType
        - payDay
        - quantityFee
        - settlementAmount
        - totalAmount
        - totalFee
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerType:
          type: string
          example: PERSONAL
        totalAmount:
          type: number
          format: double
          example: 500
        amountFee:
          type: number
          format: double
          example: 18
        quantityFee:
          type: integer
          example: 12
        totalFee:
          type: integer
          example: 12
        settlementAmount:
          type: number
          format: double
          example: 500
        payDay:
          type: string
          example: 12-03-2023
        createDate:
          type: string
          example: 12-03-2023
    CreditPayRequestDto:
      required:
        - amount
        - customerId
        - id
        - payDay
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        amount:
          type: number
          format: double
          example: 500
        payDay:
          type: string
          example: 12-03-2023
    CreditPayResponseDto:
      required:
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
...